requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.1+cpu
numpy>=1.24
Pillow>=10.0
plotly>=5.18
streamlit>=1.30
